refact: 컴포넌트, 컨테이너로 분리
