+++ client/src/Container/MakeAppointmentContainer.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import MakeAppointmentComponent from "../components/MakeAppointmentComponent";
+
+const MakeAppointmentContainer = () => {
+  return <MakeAppointmentComponent />;
+};
+
+export default MakeAppointmentContainer;
