+++ client/src/components/AdminAppointmentComponent.jsx
@@ -1,13 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
+import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import Button from "./Button";
-import { useNavigate } from "react-router-dom";
 import Modal from "./Modal";
 import SelfDiagnosisComponent from "./SelfDiagnosisComponent";
 import SendSMS from "./SendSMS";
-import axios from "axios";
 import { Container, Header, Wrapper } from "../styles/CommonStyle";
-import { toast } from "react-toastify";
 import Loader from "./Loader";
 import {
   TableHeader,
@@ -16,80 +14,23 @@ import {
   TableCell,
 } from "../styles/TableStyle";
 
-const AdminAppointmentComponent = () => {
+const AdminAppointmentComponent = ({
+  searchUsername,
+  searchDate,
+  filteredAppointments,
+  isSMSModalOpen,
+  SMSInfo,
+  appointments,
+  appointmentModals,
+  isLoading,
+  handleUsernameChange,
+  handleDateChange,
+  handleClickBillingButton,
+  handleModal,
+  handleSMSModal,
+}) => {
   const navigate = useNavigate();
-  const [searchUsername, setSearchUsername] = useState("");
-  const [searchDate, setSearchDate] = useState("");
-  const [filteredAppointments, setFilteredAppointments] = useState([]);
-  const [isSMSModalOpen, setIsSMSModalOpen] = useState(false);
-  const [SMSInfo, setSMSInfo] = useState({});
-  const [appointments, setAppointments] = useState([]);
-  const [appointmentModals, setAppointmentModals] = useState(
-    appointments.map(() => false)
-  );
-  const [isLoading, setIsLoading] = useState(false);
 
-  useEffect(() => {
-    const fetchData = async () => {
-      setIsLoading(true);
-      try {
-        const response = await axios.get("/doctor/reservation-view", {
-          headers: {
-            "Content-Type": "application/json",
-          },
-        });
-        setAppointments(response.data.data);
-        setIsLoading(false);
-      } catch (error) {
-        toast.error("데이터를 읽어오는데 실패했습니다.");
-        setIsLoading(false);
-      }
-    };
-
-    fetchData();
-  }, []);
-
-  const handleUsernameChange = (event) => {
-    setSearchUsername(event.target.value);
-    filterAppointments(event.target.value, searchDate);
-  };
-
-  const handleDateChange = (event) => {
-    setSearchDate(event.target.value);
-    filterAppointments(searchUsername, event.target.value);
-  };
-  const filterAppointments = (username, date) => {
-    const filtered = appointments.filter((appointment) => {
-      const nameMatch = username ? appointment.name.includes(username) : true;
-      const dateMatch = date
-        ? appointment.reservationDate.includes(date)
-        : true;
-      return nameMatch && dateMatch;
-    });
-
-    setFilteredAppointments(filtered);
-  };
-
-  const handleClickBillingButton = (appointment) => {
-    navigate(`/billing?id=${appointment.id}`, {
-      state: appointment,
-    });
-  };
-
-  const handleModal = (index) => {
-    const updatedModals = [...appointmentModals];
-    updatedModals[index] = !updatedModals[index];
-    setAppointmentModals(updatedModals);
-  };
-
-  const handleSMSModal = ({ appointment } = {}) => {
-    setIsSMSModalOpen(!isSMSModalOpen);
-    setSMSInfo(appointment);
-  };
-
-  useEffect(() => {
-    filterAppointments(searchUsername, searchDate);
-  }, [searchUsername, searchDate, appointments]);
   return (
     <Container>
       <Wrapper>
