+++ client/src/components/AdminWaitingListComponent.jsx
@@ -1,111 +1,17 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { styled } from "styled-components";
 import { palette } from "../styles/GlobalStyles";
-import axios from "axios";
 import { Header, Wrapper, Container } from "../styles/CommonStyle";
 import Loader from "./Loader";
 
-const AdminWaitingListComponent = () => {
-  const [appointment, setAppointment] = useState([]);
-  const currentTime = new Date();
-  const year = currentTime.getFullYear();
-  const month = String(currentTime.getMonth() + 1).padStart(2, "0");
-  const day = String(currentTime.getDate()).padStart(2, "0");
-  const today = `${year}-${month}-${day}`;
-  const sortedAppointments =
-    appointment.length >= 2
-      ? [...appointment].sort(compareAppointments)
-      : appointment;
-  const [tasks, setTasks] = useState({
-    대기중: sortedAppointments,
-    진료중: [],
-    완료: [],
-  });
-  const [isLoading, setIsLoading] = useState(false);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      setIsLoading(true);
-      try {
-        const response = await axios.get("/doctor/waiting-list-view", {
-          headers: {
-            "Access-Control-Allow-Origin": "*",
-            "Content-Type": "application/json",
-          },
-          withCredentials: true,
-        });
-        const fetchedAppointments = response.data.data;
-
-        // 정렬된 상태를 가져오기 위해 새로운 배열 생성
-        const sortedAppointments = [...fetchedAppointments].sort(
-          compareAppointments
-        );
-
-        setAppointment(sortedAppointments);
-
-        setTasks({
-          대기중: sortedAppointments,
-          진료중: [],
-          완료: [],
-        });
-        setIsLoading(false);
-      } catch (error) {
-        console.error(error);
-        setIsLoading(false);
-      }
-    };
-
-    fetchData();
-  }, []);
-
-  function compareAppointments(appointment1, appointment2) {
-    const dateComparison = appointment1.reservationDate.localeCompare(
-      appointment2.reservationDate
-    );
-
-    if (dateComparison === 0) {
-      return appointment1.reservationTime.localeCompare(
-        appointment2.reservationTime
-      );
-    }
-
-    return dateComparison;
-  }
-
-  const handleDragStart = (e, appointment, status) => {
-    e.dataTransfer.setData(
-      "text/plain",
-      JSON.stringify({ appointment, status })
-    );
-  };
-
-  const handleDragOver = (e) => {
-    e.preventDefault();
-  };
-
-  const handleDrop = (e, targetStatus) => {
-    e.preventDefault();
-    const data = e.dataTransfer.getData("text/plain");
-    const { appointment, status } = JSON.parse(data);
-
-    if (status !== targetStatus) {
-      const updatedTasks = {
-        ...tasks,
-        [status]: tasks[status].filter((apt) => apt.id !== appointment.id), // 기존 상태에서 해당 약속 제거
-        [targetStatus]: [...tasks[targetStatus], appointment],
-      };
-      setTasks(updatedTasks);
-      localStorage.setItem("tasks", JSON.stringify(updatedTasks));
-    }
-  };
-
-  useEffect(() => {
-    const storedTasks = localStorage.getItem("tasks");
-    if (storedTasks) {
-      setTasks(JSON.parse(storedTasks)); // 읽어온 값 tasks초기화
-    }
-  }, [appointment]);
-
+const AdminWaitingListComponent = ({
+  today,
+  isLoading,
+  tasks,
+  handleDragStart,
+  handleDragOver,
+  handleDrop,
+}) => {
   return (
     <Container>
       <Wrapper>
