+++ client/src/components/ChatComponent.jsx
@@ -1,23 +1,22 @@
-import React, { useState, useEffect, useRef } from "react";
-
-import { over } from "stompjs";
-import SockJS from "sockjs-client";
+import React from "react";
 import styled from "styled-components";
 import { palette } from "../styles/GlobalStyles";
-import { Container, Header, Wrapper } from "../styles/CommonStyle";
+import { Container, Wrapper } from "../styles/CommonStyle";
 
-var stompClient = null;
-
-const ChatApp = () => {
-  const [username, setUsername] = useState("");
-  const [currentMessage, setCurrentMessage] = useState("");
-  const [messages, setMessages] = useState([]);
-  const [connected, setConnected] = useState(false);
-  const [isConnecting, setIsConnecting] = useState(false);
-  const messagesContainerRef = useRef(null);
-
+const ChatComponent = ({
+  username,
+  setUsername,
+  currentMessage,
+  setCurrentMessage,
+  messages,
+  setMessages,
+  connected,
+  isConnecting,
+  connect,
+  sendMessage,
+  messagesContainerRef,
+}) => {
   var colors = [
-    // 색상 배열
     "#2196F3",
     "#32c787",
     "#00BCD4",
@@ -50,64 +49,6 @@ const ChatApp = () => {
     var index = Math.abs(hash % colors.length);
     return colors[index];
   }
-
-  const connect = () => {
-    setIsConnecting(true);
-    let socket = new SockJS("/ws/chat");
-    stompClient = over(socket);
-    stompClient.connect({}, (frame) => {
-      setConnected(true);
-      setIsConnecting(false);
-      stompClient.subscribe("/topic/public", (sdkEvent) => {
-        onMessageReceived(sdkEvent);
-      });
-      stompClient.send(
-        "/app/chat.addUser",
-        {},
-        JSON.stringify({ sender: username, type: "JOIN" })
-      );
-    });
-  };
-
-  const sendMessage = () => {
-    if (stompClient) {
-      stompClient.send(
-        "/app/chat.sendMessage",
-        {},
-        JSON.stringify({
-          sender: username,
-          type: "CHAT",
-          content: currentMessage,
-        })
-      );
-      setCurrentMessage("");
-    }
-  };
-
-  const onMessageReceived = (payload) => {
-    let message = JSON.parse(payload.body);
-    setMessages((messages) => [...messages, message]);
-    if (messagesContainerRef.current) {
-      messagesContainerRef.current.scrollTop =
-        messagesContainerRef.current.scrollHeight;
-    }
-  };
-
-  useEffect(() => {
-    return () => {
-      if (stompClient) {
-        stompClient.disconnect();
-      }
-    };
-  }, []);
-  useEffect(() => {
-    // 스크롤을 맨 아래로 이동하는 코드
-    if (messagesContainerRef.current) {
-      messagesContainerRef.current.scrollTop =
-        messagesContainerRef.current.scrollHeight;
-    }
-  }, [messages]);
-
   return (
     <Container>
       <Wrapper>
@@ -168,7 +109,7 @@ const ChatApp = () => {
   );
 };
 
-export default ChatApp;
+export default ChatComponent;
 
 const Avatar = styled.div`
   height: 35px;
