+++ client/src/pages/AdminWaitingList.jsx
@@ -2,11 +2,7 @@ import React from "react";
 import AdminWaitingListContainer from "../Container/AdminWaitingListContainer";
 
 const AdminWaitingList = () => {
-  return (
-    <div>
-      <AdminWaitingListContainer />
-    </div>
-  );
+  return <AdminWaitingListContainer />;
 };
 
 export default AdminWaitingList;
