+++ client/src/pages/Chat.jsx
@@ -2,11 +2,7 @@ import React from "react";
 import ChatContainer from "../Container/ChatContainer";
 
 const Chat = () => {
-  return (
-    <div>
-      <ChatContainer />
-    </div>
-  );
+  return <ChatContainer />;
 };
 
 export default Chat;
