+++ client/src/pages/MakeAppointment.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import MakeAppointmentContainer from "../Container/MakeAppointmentContainer";
+
+const MakeAppointment = () => {
+  return <MakeAppointmentContainer />;
+};
+
+export default MakeAppointment;
