+++ client/src/routesCollection.js
@@ -13,12 +13,11 @@ import Chat from "./pages/Chat";
 import HospitalPage from "./pages/HospitalPage";
 import AdminAppointment from "./pages/AdminAppointment";
 import AdminWaitingList from "./pages/AdminWaitingList";
-import TeleConsult from "./pages/TeleConsult";
 import Accounting from "./pages/Accounting";
 import Billing from "./pages/Billing";
-import SearchHospital from "./pages/SearchHospital";
 import MedicalCareManagement from "./pages/MedicalCareManagement";
 import KakaoAuth from "./pages/KakaoAuth";
+import MakeAppointment from "./pages/MakeAppointment";
 
 const RoutesCollection = () => {
   return (
@@ -36,7 +35,7 @@ const RoutesCollection = () => {
       <Route path={path.adminWaitingList} element={<AdminWaitingList />} />
       <Route path={path.accounting} element={<Accounting />} />
       <Route path={path.billing} element={<Billing />} />
-      <Route path={path.search} element={<SearchHospital />} />
+      <Route path={path.search} element={<MakeAppointment />} />
       <Route path={path.kakao} element={<KakaoAuth />} />
 
       <Route
